requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
future==0.16.0
idna==2.6
markovify==0.6.2
oauthlib==2.0.4
pkg-resources==0.0.0
python-twitter==3.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
Unidecode==0.4.21
urllib3==1.22
